Require Travis use a modern g++ compiler
Although Fagin has no C++ code, it imports Synder, which has to be
compiled from source. Synder requires C11, so I need a modern compiler,
which Travis doesn't provide. It is a little odd that a company whose
whole schtick is checking build success doesn't have a modern version of
the GNU Compiler, this is GCC, for godsakes. Anyway, I have to add
several lines of boilerplate to tell Travis to install a newish version
of GCC.