Remove minlength setting
This setting allowed a threshold to be placed on the minimum length of a
interval in a synteny map. Robustness analysis of Fagin results to the
density of the map is important. I will add dedicated code for this at
some point. But just slapping on a hard threshold is ugly.