Fixed the "bug"
OK, so there wasn't actually a bug in the last commit. The parse failed
because no parent could be found for an entry such as:

```
 ... ID=bar1
 ... ID=foo1;Parent=bar1
```

The problem is simply that the GFF file used "ID" and my lookup was for
"Id". Now, "ID" is the usual way to write it, so I have fixed this in
the code. However, I should probably make the comparison case
insensitive.  Also, I should make the error message spit out the
offending line as well.

In this commit, I clean up the error messages slightly and fix the "ID"
versis "Id" mismatch. I'll set up the infrastructure for more
informative error messages in the future.

First write parsers:

 1. FASTA
 2. synteny maps
 3. Synder output
 4. Newick tree

Once the individual parsers are ready:

 5. Write a record for all data on each species
 6. Write a complete Fagin data record
 7. Write filters to check it

Then derive data

 8. Write structure for coupling a [GeneModel] to a Genome
 9. Write functions for extracting model sequences
 10. Add assert that extracted data are correct

Finally:

 11. Find/build a diverse set input data, distill them to the minimal
     interesting thing, and write them into tests.
 12. Add Haddock documentation to the whole thing

The next phase of the Fagin rewrite is to:

 1. add a YAML config file and parser
 2. check existence of all files
 3. load/vet them with the above code
 4. move them all into the appropriate output folder

Then I will need to think of how to deal with the R side of fagin and
the LaTeX side that generated the report. R is powerful and succinct,
but is too sloppy. I'll need to atomize the R codebase and wrap each
atom in a Haskell function that can call it. There is a program for
interfacing Haskell with R using quasiquotation (I don't really know
that that is yet). This may be the best choice.

As for generating the report. I think I will add a Result monad into the
mix, which accumulates results as they appear. Perhaps creating a
Markdown file. I need to learn about Monad transformers. But not yet,
first finish phase 1 and 2.