Add show instances and model2gff function
A goal of the the Haskell side of Fagin is to prepar a vetted dataset
from the biologist prepared inputs. So a common theme will be reading in
a dirty data set, building it into a datastructure with solid
validation, then writing it back to the format it originaly came from.

I am way too tired to be writing, I'll update this commit tomorrow