House cleaning and profiler setup
There are a bunch of little, inconsequential changes here. The main ones
are that I added profiling and a larger data set.

I also found the source of my performance problem. The `splitSeq`
functions are fantastically slow. The function that splits on tabs
required over a third of the time for the entire program. I replaced
this with `words`, which was so fast it didn't even register as taking
time. Now, `words` doesn't quite do what I need (it splits on any
whitespace), but the TAB-splitter should be at least as fast. Once I
replace all the `splitSeq` functions I should be 10X faster, or so.