Convert back to ShortByteString
Originally the strings I parsed from the GFF where stored as slices
referencing the entire ByteString slurped into memory from the GFF file.
These references held the entire thing in memory, long past when it was
needed. A better solution is to copy the data. I could copy into
ByteStrings, which would circumvent to need to map back and forth
between ShortByteString and ByteString. Basically, the choice between
the two is a tradeoff between time and space. I go with space.

The result? Now the GFF file is cleaned out of memory after the readGff
function finishes with it. This yields a nice 10% drop or so in maximum
running memory.

I don't actually have a clear way of quantifying maximum memory
currently. But the memory plots I get from GHC are beautiful. It would
be nice to have a diff benchmark that compared between commits.