Add data classes
I will generally endeavour to not store anything big in memory unless
necessary, and then only temporarily. So rather than storing, say, a
whole genome, I will open it once, check it for correctness and extract
summaries (small enough to be stored in memory) and then save it as an R
data object (for fast loading later). Then I will have getter methods
that retrieve the full file on demand. This will also allow me to later
refactor to store the data in a database or anything else I want,
without the user needing to know.

Now, I don't want to implement getter and setter methods for every slot
in every S4 class. Most of these classes are implemented to
transparently store a dataset, all of which should be visible to the
user. Also, setter methods will not need to be used, for the most part,
since most the data should be immutable.

Hmm, to write getters or not to write getters ...