Add ORF extractor
 ** SILENT BUG CREATED **

I am simply extracting the START STOP bound intervals with a simple
regular expression. This is not the right way to do it. Also, the
regular expressions are non-overlapping, which means they will find the
first ATG, then read to the nearest stop. Then resume searching for the
next ATG after the stop. This is completely the wrong.