Fixed all errors in yeast-based pipeline data prep
Rsamtools/GenomicFeatures are not quite the panaceae I had hoped they
would be. GenomicFeatures subset sequences extremely slowly if there are
many sequences in the genome. It is blazing fast if there are few. So
now there is just one step that is slow,
`GenomicFeatures::extractTranscriptSeqs`.

I also need to rewire the caching system in fagin, since the txdb and
indexed fasta data can't be stored as Rdata (or at least should not be).