Debugged DNA to genome aligments
This is the last piece of the secondary data. Now, with a bit of
rewiring, I can build the feature table and then the final
classifications.

Here is an updated TODO:

 1. Fix the issue in rmonad that breaks somewhere in the secondary code

 2. Think of what to do with echos

 3. Storage and caching (rmonad should handle caching)

 4. Finer rmonad granularity in alignment functions

 5. feature table and classification

rmonad is still proving difficult to work with. I need more utility
functions for extracting nodes, getting last valid input, and the like.