Return to name based in cdsBy and exonsBy
The ID field in the TxDb objects in an integer. The value of the ID is
stored in the name (I need to double check this). I was confusing this
early, thinking that the name was related to the Name field commonly
used in the attribute column of GFF files.

I am still a little uncertain what exactly TxDb is doing, how exactly it
goes from GFF to sqlite database, but at least with the yeast data, it
is storing the ID names as tx_name and storing an intege as tx_id.