Refactor secondary data
There is much that could be done here that is not. I had written an echo
that I did not use. Synder is not used to the fullest extent. Also much
of the intermediate data is not preserved (for example the identifies of
the genes that are matched). I should cache all of this. I should also
write dedicated classes for each of the data sets.

I need to walk through and carefully re-evaluate each step of the
precedure.