Add rough Excel report
The images are really shoddy. Currently I make PNGs and import them into
the excel file. However, this is awful for two reasons: 1) the
resolution and scaling are piss poor, 2) they are not interactive.

I really need to get some sort of vector graphic that can be interacted
with in the excel context (e.g. change title and axis names).

I also need a little more formatting on the worksheets. I could put all
the short summary tables on one page, for example.

Also I need to collate the mammoth table, with all the genewise data.