Possibly resolved analysis 1
My solution to the synder loop is monstrous.

I have to pass the seqinfo and species names, for both the target and
query species. The the `genome` field that holds the species name is
automatically deleted when a Seqinfo object is assigned to a sequence.
They search for specific information in their genome browsers, if the
species isn't in the browser, they reannotate it as "unspecified".

Why is there such a special link between a supposedly general
bioinformatics package and one specific genome browser? I hope the
janitor at USCS spills mop water all over their servers.