I want cake
My tests are failing due to a bug in ORFik. I issued a bug report to the
Norsemen that maintain that codebase. It should be resolved soon.

My pipeline also breaks due to a bug in rmonad -- a most tricksy bug
hiding somewhere deep in the non-standard evaluation quagmire that is
rmonad. Lack-a-day. I commented out the section in `fagin` that used the
broken `rmonad` feature. But now I get 6037 warnings. 6037 warnings that
rmonad would compress into one if only `rmonad` weren't broken.

I am sad. I want cake.