determine_origins is working
Next step is to run all of the yeast species through and very carefully
study the results.

There seems to be a mistake in the early data preparation. Currently
almost exactly half of the controls are missing gene matches but have
CDS matches. This suggests that the translations on the reverse strand
are incorrect.