Fix bug skipped minus strand bug in overlapMap
By default, the `GenomicRanges::findOverlaps` function is strand
sensitive. However, the search intervals `synder` produces are
independent of strand (they are all arbitrarily set to '+'). This means
that all features on the '-' strand are skipped in the overlap map. This
resulted in about half of the control genes (ancient proteins) having no
identified protein homolog. The fix for the problem is simple to set
`ignore.strand=TRUE` in `findOverlaps`.