Use Saccharomyces-specific genes (not orphans)
Previously I used species-specific genes as queries in fagin. This is
silly, since the coolest use of fagin is to find positive evidence of
the lineage-specific de novo events. So it makes more sense to use
Saccharomyces specific genes as the "queries" and genes from older
strata as the controls. I use the phylostratigraphy results from
`phylostratr`.