Workaround rmonad bug
There is currently a bug in rmonad (see arendsee/rmonad#19) that drops
nodes when chaining `%__%` rmonad operators. In `x %>>% y %>>% z`, the
`y` node is lost. A workaround is to use the nearly equivalent rmonad
pattern `x %>>% {y} %>>% {z}`. The two expressions are not quite the
same, since in the workaround a value is passed (but not used) from the
left to the right (e.g. x is passed to `{y}` but ignored). The
workaround *should* preserve the logic of the original at a loss of some
elegance.