Remove bad adjustment
For reasons I can attribute only to carelessness, I was adjusting for
multiple testing by correcting for the number of query genes. This is a
completely non-sensical thing to do and would make the significance of a
hit dependent on the number of genes I am searching.

I get a p-value for the significance of each query against each target
sequence. For aa-vs-aa searches, this means for a given query I have one
p-value for each aa that overlaps a search interval.

Probably the most non-controversial approach would be to adjust for the
total number of sequences the query is searched against. This would
emulate what BLAST does when all species are pooled into one database.