Add correct adjustment
Now p-values are adjusted the "right" way, as described in the prior
commit. The final adjusted p-values are a function of the number of
sequences searched across all species, so it was more convenient to
change the tags to be "<tag>/<group>/<species>", which made is easy to
pull out all the data for a given tag and group (e.g., "aa2aa/query").
I also changed the data stored for "gene2genome", making the DNA to DNA
search better match all the AA to AA searches. This required a bit of
refactoring, but the yeast case study at least appears to be running as
expected, so nothing is terribly broken ...